EnemiesSpawner && ObstacleGeneration
Added functions to Obstacle Generation to be able to activate, deactivate, and shuffle the obstacle floor.

Modified the Enemies Spawner so that it can spawn enemies with given params and check for empty location before spawning them. If there's no valid location to spawn enemies then NO enemies will be spawned.